backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
transformers
torch
pydantic
python-dotenv
requests